add robustness to generated loop code
Certain expressions--integers (with or without a preceding +/-), null,
and void--when used as sources of for loops resulted in illegal
JavaScript being generated. While (42).length is a silly but legal thing
to do, 42.length is an actual SyntaxError in JS. This commit adds checks
to wrap potentially problematic expressions in parentheses.