Fix for gkz/LiveScript#926
This issue was caused by a failure of `traverse-children` to enter the
topic of a `match` statement (it worked fine with `switch`). This in
turn was because the Switch AST node, when representing a match, stored
its thing-to-match in a member called `target` instead of `topic`, but
'target' was not included in its list of child fields ('topic' was).

I couldn't see any use in keeping two distinct `topic` and `target`
fields, since they're never both used in the same node and never
referenced from other classes, so the issue was fixed by unifying them.