refactor object syntax, part 1
This commit is a refactoring that doesn't touch the language grammar (it
does touch the interface between the grammar and AST). It prepares for
the grammar refactoring that comes next. No language semantics are
changed.

The big idea here is to have Objs hold only Props (well, and comments), and
to pull out any and/or/? logic onto the Prop instead of having it wrap
the Prop or the value inside the Prop. Splats in Objs are now
Prop(Splat!, contents), and keyless properties are simply Props with a
null key.

The refactoring also pulls out slice expanding and property shorthand
expanding as a distinct pass from the rest of the compilation; this
makes them easier to reason about and cleans up the rest of the code a
little.