bodyless let
This commit adds two language features involving the `let` keyword. The
first is bodyless let, which permits a `let` line without an indented
body to be interpreted as implicitly containing the remainder of the
current block as its body, much like backcall syntax. The second is `let
const` and `let var`, which are shorthands for including const or var
declarations as the first line in the body of a let expression, whether
that body is indented or implicit.

Resolve gkz/LiveScript#962.