stop embedding source map with -ce
68a6d21 introduced a regression in the functionality of 9d32954 by
causing a source map to be embedded by default when using the -ce
combination of compiler flags, in contradiction with the comment of
9d32954 which states that embedded source maps should only be the
default when the compiler is running, not just compiling, code. With
68a6d21, the presence of `o.eval` no longer forces the compiler to run
code even if -c is used, so `o.eval` is no longer relevant to the
default value of `o.map`.