show parser build errors, make them fail the build
This makes build-parser print error messages about shift/reduce and
reduce/reduce conflicts to stderr instead of prepending them to
lib/parser.js. It also halts the build and leaves lib/parser.js
untouched if there are any such messages. It's a bit of a hack, but it
greatly improves the ergonomics of tinkering with the grammar.