fix gkz/LiveScript#1098
Most of the time, the precedence difference between `&&` and `||`
doesn't need to be attended, because it's the same in LiveScript and
JavaScript. But when synthetic `&&` nodes are created, it's important to
ensure that the children of the node either must have had higher
precedence when parsed, are also synthetic and represent expressions
with higher precedence, or are wrapped in `Parens`. The reported bug,
and a similar issue with the binary `?` operator, are the consequences
if not.