fix gkz/LiveScript#1102
This issue was a regression from 1.5.0 first appearing in commit
8cffc57, in which I introduced the `rewrite-shorthand` method and moved
slice expanding into it. The problem with my refactoring is that it
caused everything in `rewrite-shorthand` to happen before `unfold-soak`,
but the code in the issue relies on `unfold-soak` happening prior to
slices being expanded.

This commit prepends a short-circuiting call to `unfold-soak` in the
`rewrite-shorthand` method of every node type that defines a non-trivial
`unfold-soak`.